Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -2,7 +2,7 @@
 #include<stdio.h>
 int main()
 {
-    float x;
+    double x;
     scanf("%lf",&x);
-    printf("You entered: %d",x);
+    printf("You entered: %lf",x);
 }
